# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: DeFarm Engines API
  version: 1.0.0
  description: |
    DeFarm Engines API provides a comprehensive agricultural supply chain tracking system
    with tokenization, circuit management, and secure item verification.

    ## Key Features
    - **JWT Authentication** - Secure token-based authentication
    - **Circuit Management** - Permission-controlled data sharing repositories
    - **Item Tokenization** - Local ID (LID) to DFID conversion via circuits
    - **Enhanced Identifiers** - Canonical and contextual identifier support
    - **Storage Adapters** - IPFS, Stellar blockchain, and custom storage backends
    - **Webhook System** - Post-action webhooks for circuit events
    - **Audit Trail** - Complete event history and activity tracking

    ## Authentication
    All authenticated endpoints require a JWT token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```

    Or API key authentication:
    ```
    X-API-Key: dfm_<your-api-key>
    ```
  contact:
    name: DeFarm API Support
    email: api@defarm.net
  license:
    name: Proprietary
servers:
  - url: https://connect.defarm.net/api
    description: Production server
  - url: http://localhost:3000/api
    description: Local development server
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login

    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key in format dfm_<32-character-string>
  schemas:
...
    Identifier:
      type: object
      required:
        - key
        - value
      properties:
        namespace:
          type: string
          description: Logical namespace (bovino, cafe, soja, etc.)
          example: "bovino"
        key:
          type: string
          example: "sisbov"
        value:
          type: string
          example: "BR12345678901234"
        id_type:
          type: string
          enum: [Canonical, Contextual]
          description: |
            - Canonical: Globally unique (SISBOV, CPF, etc.)
            - Contextual: Scoped identifier (requires scope)
          default: Contextual
        scope:
          type: string
          nullable: true
          description: Scope for contextual identifiers (user, organization, circuit)
          example: "user"
        verified:
          type: boolean
          nullable: true
          description: Verification flag for canonical identifiers
        verification_date:
          type: string
          format: date-time
          nullable: true
          description: When the canonical identifier was verified
...
    ReceiptResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Receipt ID
        hash:
          type: string
          description: BLAKE3 hash of the data
          example: "abc123..."
        timestamp:
          type: integer
          format: int64
          description: Unix timestamp when receipt was created
        data_size:
          type: integer
          description: Size of original data in bytes
        identifiers:
          type: array
          items:
            $ref: '#/components/schemas/Identifier'
...
paths:
...
  /api-keys/{key_id}:
    delete:
      tags:
        - API Keys
      summary: Delete API key
      description: Permanently delete an API key
      operationId: deleteApiKey
      parameters:
        - name: key_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: API key deleted successfully

  # ============================================================================
  # Event & Activity Endpoints
  # ============================================================================
...
  /receipts/search/key/{key}:
    get:
      tags:
        - Receipts
      summary: Search receipts by identifier key
      description: Find all receipts containing identifiers with this key
      operationId: searchByKey
      security: []
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of matching receipts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReceiptResponse'
...